TASK: Add buildstack and asset forwarding
