--- .styleci.yml
@@ -1,6 +1,0 @@
-preset: psr2
-
-finder:
-  path:
-    - "Classes"
-    - "Tests"
