+++ Resources/Private/SetupDashboard/tailwind.config.js
@@ -1,0 +1,8 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+	content: ["./src/**/*.{ts,html}"],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
