+++ Resources/Public/SetupDashboard/main.js
@@ -1,0 +1,6 @@
+(()=>{var Ie=!1,$e=!1,P=[],Pe=-1;function Lr(e){Fr(e)}function Fr(e){P.includes(e)||P.push(e),Kr()}function wt(e){let t=P.indexOf(e);t!==-1&&t>Pe&&P.splice(t,1)}function Kr(){!$e&&!Ie&&(Ie=!0,queueMicrotask(Dr))}function Dr(){Ie=!1,$e=!0;for(let e=0;e<P.length;e++)P[e](),Pe=e;P.length=0,Pe=-1,$e=!1}var H,k,Q,Et,Ne=!0;function Br(e){Ne=!1,e(),Ne=!0}function Hr(e){H=e.reactive,Q=e.release,k=t=>e.effect(t,{scheduler:r=>{Ne?Lr(r):r()}}),Et=e.raw}function _t(e){k=e}function kr(e){let t=()=>{};return[n=>{let i=k(n);return e._x_effects||(e._x_effects=new Set,e._x_runEffects=()=>{e._x_effects.forEach(o=>o())}),e._x_effects.add(i),t=()=>{i!==void 0&&(e._x_effects.delete(i),Q(i))},i},()=>{t()}]}var St=[],At=[],Ot=[];function zr(e){Ot.push(e)}function Ct(e,t){typeof t=="function"?(e._x_cleanups||(e._x_cleanups=[]),e._x_cleanups.push(t)):(t=e,At.push(t))}function qr(e){St.push(e)}function Wr(e,t,r){e._x_attributeCleanups||(e._x_attributeCleanups={}),e._x_attributeCleanups[t]||(e._x_attributeCleanups[t]=[]),e._x_attributeCleanups[t].push(r)}function Mt(e,t){e._x_attributeCleanups&&Object.entries(e._x_attributeCleanups).forEach(([r,n])=>{(t===void 0||t.includes(r))&&(n.forEach(i=>i()),delete e._x_attributeCleanups[r])})}var Je=new MutationObserver(Qe),Ge=!1;function Ye(){Je.observe(document,{subtree:!0,childList:!0,attributes:!0,attributeOldValue:!0}),Ge=!0}function Tt(){Ur(),Je.disconnect(),Ge=!1}var V=[],Ce=!1;function Ur(){V=V.concat(Je.takeRecords()),V.length&&!Ce&&(Ce=!0,queueMicrotask(()=>{Vr(),Ce=!1}))}function Vr(){Qe(V),V.length=0}function x(e){if(!Ge)return e();Tt();let t=e();return Ye(),t}var Ze=!1,ae=[];function Jr(){Ze=!0}function Gr(){Ze=!1,Qe(ae),ae=[]}function Qe(e){if(Ze){ae=ae.concat(e);return}let t=[],r=[],n=new Map,i=new Map;for(let o=0;o<e.length;o++)if(!e[o].target._x_ignoreMutationObserver&&(e[o].type==="childList"&&(e[o].addedNodes.forEach(s=>s.nodeType===1&&t.push(s)),e[o].removedNodes.forEach(s=>s.nodeType===1&&r.push(s))),e[o].type==="attributes")){let s=e[o].target,a=e[o].attributeName,u=e[o].oldValue,c=()=>{n.has(s)||n.set(s,[]),n.get(s).push({name:a,value:s.getAttribute(a)})},l=()=>{i.has(s)||i.set(s,[]),i.get(s).push(a)};s.hasAttribute(a)&&u===null?c():s.hasAttribute(a)?(l(),c()):l()}i.forEach((o,s)=>{Mt(s,o)}),n.forEach((o,s)=>{St.forEach(a=>a(s,o))});for(let o of r)if(!t.includes(o)&&(At.forEach(s=>s(o)),o._x_cleanups))for(;o._x_cleanups.length;)o._x_cleanups.pop()();t.forEach(o=>{o._x_ignoreSelf=!0,o._x_ignore=!0});for(let o of t)r.includes(o)||o.isConnected&&(delete o._x_ignoreSelf,delete o._x_ignore,Ot.forEach(s=>s(o)),o._x_ignore=!0,o._x_ignoreSelf=!0);t.forEach(o=>{delete o._x_ignoreSelf,delete o._x_ignore}),t=null,r=null,n=null,i=null}function Rt(e){return ee(K(e))}function X(e,t,r){return e._x_dataStack=[t,...K(r||e)],()=>{e._x_dataStack=e._x_dataStack.filter(n=>n!==t)}}function ht(e,t){let r=e._x_dataStack[0];Object.entries(t).forEach(([n,i])=>{r[n]=i})}function K(e){return e._x_dataStack?e._x_dataStack:typeof ShadowRoot=="function"&&e instanceof ShadowRoot?K(e.host):e.parentNode?K(e.parentNode):[]}function ee(e){let t=new Proxy({},{ownKeys:()=>Array.from(new Set(e.flatMap(r=>Object.keys(r)))),has:(r,n)=>e.some(i=>i.hasOwnProperty(n)),get:(r,n)=>(e.find(i=>{if(i.hasOwnProperty(n)){let o=Object.getOwnPropertyDescriptor(i,n);if(o.get&&o.get._x_alreadyBound||o.set&&o.set._x_alreadyBound)return!0;if((o.get||o.set)&&o.enumerable){let s=o.get,a=o.set,u=o;s=s&&s.bind(t),a=a&&a.bind(t),s&&(s._x_alreadyBound=!0),a&&(a._x_alreadyBound=!0),Object.defineProperty(i,n,{...u,get:s,set:a})}return!0}return!1})||{})[n],set:(r,n,i)=>{let o=e.find(s=>s.hasOwnProperty(n));return o?o[n]=i:e[e.length-1][n]=i,!0}});return t}function It(e){let t=n=>typeof n=="object"&&!Array.isArray(n)&&n!==null,r=(n,i="")=>{Object.entries(Object.getOwnPropertyDescriptors(n)).forEach(([o,{value:s,enumerable:a}])=>{if(a===!1||s===void 0)return;let u=i===""?o:`${i}.${o}`;typeof s=="object"&&s!==null&&s._x_interceptor?n[o]=s.initialize(e,u,o):t(s)&&s!==n&&!(s instanceof Element)&&r(s,u)})};return r(e)}function $t(e,t=()=>{}){let r={initialValue:void 0,_x_interceptor:!0,initialize(n,i,o){return e(this.initialValue,()=>Yr(n,i),s=>je(n,i,s),i,o)}};return t(r),n=>{if(typeof n=="object"&&n!==null&&n._x_interceptor){let i=r.initialize.bind(r);r.initialize=(o,s,a)=>{let u=n.initialize(o,s,a);return r.initialValue=u,i(o,s,a)}}else r.initialValue=n;return r}}function Yr(e,t){return t.split(".").reduce((r,n)=>r[n],e)}function je(e,t,r){if(typeof t=="string"&&(t=t.split(".")),t.length===1)e[t[0]]=r;else{if(t.length===0)throw error;return e[t[0]]||(e[t[0]]={}),je(e[t[0]],t.slice(1),r)}}var Pt={};function S(e,t){Pt[e]=t}function Le(e,t){return Object.entries(Pt).forEach(([r,n])=>{Object.defineProperty(e,`$${r}`,{get(){let[i,o]=Kt(t);return i={interceptor:$t,...i},Ct(t,o),n(t,i)},enumerable:!1})}),e}function Zr(e,t,r,...n){try{return r(...n)}catch(i){Y(i,e,t)}}function Y(e,t,r=void 0){Object.assign(e,{el:t,expression:r}),console.warn(`Alpine Expression Error: ${e.message}
+
+${r?'Expression: "'+r+`"
+
+`:""}`,t),setTimeout(()=>{throw e},0)}var se=!0;function Qr(e){let t=se;se=!1,e(),se=t}function F(e,t,r={}){let n;return b(e,t)(i=>n=i,r),n}function b(...e){return Nt(...e)}var Nt=jt;function Xr(e){Nt=e}function jt(e,t){let r={};Le(r,e);let n=[r,...K(e)],i=typeof t=="function"?en(n,t):rn(n,t,e);return Zr.bind(null,e,t,i)}function en(e,t){return(r=()=>{},{scope:n={},params:i=[]}={})=>{let o=t.apply(ee([n,...e]),i);ue(r,o)}}var Me={};function tn(e,t){if(Me[e])return Me[e];let r=Object.getPrototypeOf(async function(){}).constructor,n=/^[\n\s]*if.*\(.*\)/.test(e)||/^(let|const)\s/.test(e)?`(async()=>{ ${e} })()`:e,o=(()=>{try{return new r(["__self","scope"],`with (scope) { __self.result = ${n} }; __self.finished = true; return __self.result;`)}catch(s){return Y(s,t,e),Promise.resolve()}})();return Me[e]=o,o}function rn(e,t,r){let n=tn(t,r);return(i=()=>{},{scope:o={},params:s=[]}={})=>{n.result=void 0,n.finished=!1;let a=ee([o,...e]);if(typeof n=="function"){let u=n(n,a).catch(c=>Y(c,r,t));n.finished?(ue(i,n.result,a,s,r),n.result=void 0):u.then(c=>{ue(i,c,a,s,r)}).catch(c=>Y(c,r,t)).finally(()=>n.result=void 0)}}}function ue(e,t,r,n,i){if(se&&typeof t=="function"){let o=t.apply(r,n);o instanceof Promise?o.then(s=>ue(e,s,r,n)).catch(s=>Y(s,i,t)):e(o)}else typeof t=="object"&&t instanceof Promise?t.then(o=>e(o)):e(t)}var Xe="x-";function z(e=""){return Xe+e}function nn(e){Xe=e}var Fe={};function g(e,t){return Fe[e]=t,{before(r){if(!Fe[r]){console.warn("Cannot find directive `${directive}`. `${name}` will use the default order of execution");return}let n=$.indexOf(r);$.splice(n>=0?n:$.indexOf("DEFAULT"),0,e)}}}function et(e,t,r){if(t=Array.from(t),e._x_virtualDirectives){let o=Object.entries(e._x_virtualDirectives).map(([a,u])=>({name:a,value:u})),s=Lt(o);o=o.map(a=>s.find(u=>u.name===a.name)?{name:`x-bind:${a.name}`,value:`"${a.value}"`}:a),t=t.concat(o)}let n={};return t.map(Ht((o,s)=>n[o]=s)).filter(zt).map(an(n,r)).sort(un).map(o=>sn(e,o))}function Lt(e){return Array.from(e).map(Ht()).filter(t=>!zt(t))}var Ke=!1,U=new Map,Ft=Symbol();function on(e){Ke=!0;let t=Symbol();Ft=t,U.set(t,[]);let r=()=>{for(;U.get(t).length;)U.get(t).shift()();U.delete(t)},n=()=>{Ke=!1,r()};e(r),n()}function Kt(e){let t=[],r=a=>t.push(a),[n,i]=kr(e);return t.push(i),[{Alpine:re,effect:n,cleanup:r,evaluateLater:b.bind(b,e),evaluate:F.bind(F,e)},()=>t.forEach(a=>a())]}function sn(e,t){let r=()=>{},n=Fe[t.type]||r,[i,o]=Kt(e);Wr(e,t.original,o);let s=()=>{e._x_ignore||e._x_ignoreSelf||(n.inline&&n.inline(e,t,i),n=n.bind(n,e,t,i),Ke?U.get(Ft).push(n):n())};return s.runCleanups=o,s}var Dt=(e,t)=>({name:r,value:n})=>(r.startsWith(e)&&(r=r.replace(e,t)),{name:r,value:n}),Bt=e=>e;function Ht(e=()=>{}){return({name:t,value:r})=>{let{name:n,value:i}=kt.reduce((o,s)=>s(o),{name:t,value:r});return n!==t&&e(n,t),{name:n,value:i}}}var kt=[];function tt(e){kt.push(e)}function zt({name:e}){return qt().test(e)}var qt=()=>new RegExp(`^${Xe}([^:^.]+)\\b`);function an(e,t){return({name:r,value:n})=>{let i=r.match(qt()),o=r.match(/:([a-zA-Z0-9\-:]+)/),s=r.match(/\.[^.\]]+(?=[^\]]*$)/g)||[],a=t||e[r]||r;return{type:i?i[1]:null,value:o?o[1]:null,modifiers:s.map(u=>u.replace(".","")),expression:n,original:a}}}var De="DEFAULT",$=["ignore","ref","data","id","bind","init","for","model","modelable","transition","show","if",De,"teleport"];function un(e,t){let r=$.indexOf(e.type)===-1?De:e.type,n=$.indexOf(t.type)===-1?De:t.type;return $.indexOf(r)-$.indexOf(n)}function J(e,t,r={}){e.dispatchEvent(new CustomEvent(t,{detail:r,bubbles:!0,composed:!0,cancelable:!0}))}function M(e,t){if(typeof ShadowRoot=="function"&&e instanceof ShadowRoot){Array.from(e.children).forEach(i=>M(i,t));return}let r=!1;if(t(e,()=>r=!0),r)return;let n=e.firstElementChild;for(;n;)M(n,t,!1),n=n.nextElementSibling}function D(e,...t){console.warn(`Alpine Warning: ${e}`,...t)}function cn(){document.body||D("Unable to initialize. Trying to load Alpine before `<body>` is available. Did you forget to add `defer` in Alpine's `<script>` tag?"),J(document,"alpine:init"),J(document,"alpine:initializing"),Ye(),zr(t=>T(t,M)),Ct(t=>Zt(t)),qr((t,r)=>{et(t,r).forEach(n=>n())});let e=t=>!fe(t.parentElement,!0);Array.from(document.querySelectorAll(Vt())).filter(e).forEach(t=>{T(t)}),J(document,"alpine:initialized")}var rt=[],Wt=[];function Ut(){return rt.map(e=>e())}function Vt(){return rt.concat(Wt).map(e=>e())}function Jt(e){rt.push(e)}function Gt(e){Wt.push(e)}function fe(e,t=!1){return de(e,r=>{if((t?Vt():Ut()).some(i=>r.matches(i)))return!0})}function de(e,t){if(e){if(t(e))return e;if(e._x_teleportBack&&(e=e._x_teleportBack),!!e.parentElement)return de(e.parentElement,t)}}function ln(e){return Ut().some(t=>e.matches(t))}var Yt=[];function fn(e){Yt.push(e)}function T(e,t=M,r=()=>{}){on(()=>{t(e,(n,i)=>{r(n,i),Yt.forEach(o=>o(n,i)),et(n,n.attributes).forEach(o=>o()),n._x_ignore&&i()})})}function Zt(e){M(e,t=>Mt(t))}var Be=[],nt=!1;function it(e=()=>{}){return queueMicrotask(()=>{nt||setTimeout(()=>{He()})}),new Promise(t=>{Be.push(()=>{e(),t()})})}function He(){for(nt=!1;Be.length;)Be.shift()()}function dn(){nt=!0}function ot(e,t){return Array.isArray(t)?gt(e,t.join(" ")):typeof t=="object"&&t!==null?pn(e,t):typeof t=="function"?ot(e,t()):gt(e,t)}function gt(e,t){let r=o=>o.split(" ").filter(Boolean),n=o=>o.split(" ").filter(s=>!e.classList.contains(s)).filter(Boolean),i=o=>(e.classList.add(...o),()=>{e.classList.remove(...o)});return t=t===!0?t="":t||"",i(n(t))}function pn(e,t){let r=a=>a.split(" ").filter(Boolean),n=Object.entries(t).flatMap(([a,u])=>u?r(a):!1).filter(Boolean),i=Object.entries(t).flatMap(([a,u])=>u?!1:r(a)).filter(Boolean),o=[],s=[];return i.forEach(a=>{e.classList.contains(a)&&(e.classList.remove(a),s.push(a))}),n.forEach(a=>{e.classList.contains(a)||(e.classList.add(a),o.push(a))}),()=>{s.forEach(a=>e.classList.add(a)),o.forEach(a=>e.classList.remove(a))}}function pe(e,t){return typeof t=="object"&&t!==null?_n(e,t):hn(e,t)}function _n(e,t){let r={};return Object.entries(t).forEach(([n,i])=>{r[n]=e.style[n],n.startsWith("--")||(n=gn(n)),e.style.setProperty(n,i)}),setTimeout(()=>{e.style.length===0&&e.removeAttribute("style")}),()=>{pe(e,r)}}function hn(e,t){let r=e.getAttribute("style",t);return e.setAttribute("style",t),()=>{e.setAttribute("style",r||"")}}function gn(e){return e.replace(/([a-z])([A-Z])/g,"$1-$2").toLowerCase()}function ke(e,t=()=>{}){let r=!1;return function(){r?t.apply(this,arguments):(r=!0,e.apply(this,arguments))}}g("transition",(e,{value:t,modifiers:r,expression:n},{evaluate:i})=>{typeof n=="function"&&(n=i(n)),n?vn(e,n,t):yn(e,r,t)});function vn(e,t,r){Qt(e,ot,""),{enter:i=>{e._x_transition.enter.during=i},"enter-start":i=>{e._x_transition.enter.start=i},"enter-end":i=>{e._x_transition.enter.end=i},leave:i=>{e._x_transition.leave.during=i},"leave-start":i=>{e._x_transition.leave.start=i},"leave-end":i=>{e._x_transition.leave.end=i}}[r](t)}function yn(e,t,r){Qt(e,pe);let n=!t.includes("in")&&!t.includes("out")&&!r,i=n||t.includes("in")||["enter"].includes(r),o=n||t.includes("out")||["leave"].includes(r);t.includes("in")&&!n&&(t=t.filter((_,v)=>v<t.indexOf("out"))),t.includes("out")&&!n&&(t=t.filter((_,v)=>v>t.indexOf("out")));let s=!t.includes("opacity")&&!t.includes("scale"),a=s||t.includes("opacity"),u=s||t.includes("scale"),c=a?0:1,l=u?q(t,"scale",95)/100:1,d=q(t,"delay",0),p=q(t,"origin","center"),y="opacity, transform",O=q(t,"duration",150)/1e3,ne=q(t,"duration",75)/1e3,f="cubic-bezier(0.4, 0.0, 0.2, 1)";i&&(e._x_transition.enter.during={transformOrigin:p,transitionDelay:d,transitionProperty:y,transitionDuration:`${O}s`,transitionTimingFunction:f},e._x_transition.enter.start={opacity:c,transform:`scale(${l})`},e._x_transition.enter.end={opacity:1,transform:"scale(1)"}),o&&(e._x_transition.leave.during={transformOrigin:p,transitionDelay:d,transitionProperty:y,transitionDuration:`${ne}s`,transitionTimingFunction:f},e._x_transition.leave.start={opacity:1,transform:"scale(1)"},e._x_transition.leave.end={opacity:c,transform:`scale(${l})`})}function Qt(e,t,r={}){e._x_transition||(e._x_transition={enter:{during:r,start:r,end:r},leave:{during:r,start:r,end:r},in(n=()=>{},i=()=>{}){ze(e,t,{during:this.enter.during,start:this.enter.start,end:this.enter.end},n,i)},out(n=()=>{},i=()=>{}){ze(e,t,{during:this.leave.during,start:this.leave.start,end:this.leave.end},n,i)}})}window.Element.prototype._x_toggleAndCascadeWithTransitions=function(e,t,r,n){let i=document.visibilityState==="visible"?requestAnimationFrame:setTimeout,o=()=>i(r);if(t){e._x_transition&&(e._x_transition.enter||e._x_transition.leave)?e._x_transition.enter&&(Object.entries(e._x_transition.enter.during).length||Object.entries(e._x_transition.enter.start).length||Object.entries(e._x_transition.enter.end).length)?e._x_transition.in(r):o():e._x_transition?e._x_transition.in(r):o();return}e._x_hidePromise=e._x_transition?new Promise((s,a)=>{e._x_transition.out(()=>{},()=>s(n)),e._x_transitioning.beforeCancel(()=>a({isFromCancelledTransition:!0}))}):Promise.resolve(n),queueMicrotask(()=>{let s=Xt(e);s?(s._x_hideChildren||(s._x_hideChildren=[]),s._x_hideChildren.push(e)):i(()=>{let a=u=>{let c=Promise.all([u._x_hidePromise,...(u._x_hideChildren||[]).map(a)]).then(([l])=>l());return delete u._x_hidePromise,delete u._x_hideChildren,c};a(e).catch(u=>{if(!u.isFromCancelledTransition)throw u})})})};function Xt(e){let t=e.parentNode;if(t)return t._x_hidePromise?t:Xt(t)}function ze(e,t,{during:r,start:n,end:i}={},o=()=>{},s=()=>{}){if(e._x_transitioning&&e._x_transitioning.cancel(),Object.keys(r).length===0&&Object.keys(n).length===0&&Object.keys(i).length===0){o(),s();return}let a,u,c;xn(e,{start(){a=t(e,n)},during(){u=t(e,r)},before:o,end(){a(),c=t(e,i)},after:s,cleanup(){u(),c()}})}function xn(e,t){let r,n,i,o=ke(()=>{x(()=>{r=!0,n||t.before(),i||(t.end(),He()),t.after(),e.isConnected&&t.cleanup(),delete e._x_transitioning})});e._x_transitioning={beforeCancels:[],beforeCancel(s){this.beforeCancels.push(s)},cancel:ke(function(){for(;this.beforeCancels.length;)this.beforeCancels.shift()();o()}),finish:o},x(()=>{t.start(),t.during()}),dn(),requestAnimationFrame(()=>{if(r)return;let s=Number(getComputedStyle(e).transitionDuration.replace(/,.*/,"").replace("s",""))*1e3,a=Number(getComputedStyle(e).transitionDelay.replace(/,.*/,"").replace("s",""))*1e3;s===0&&(s=Number(getComputedStyle(e).animationDuration.replace("s",""))*1e3),x(()=>{t.before()}),n=!0,requestAnimationFrame(()=>{r||(x(()=>{t.end()}),He(),setTimeout(e._x_transitioning.finish,s+a),i=!0)})})}function q(e,t,r){if(e.indexOf(t)===-1)return r;let n=e[e.indexOf(t)+1];if(!n||t==="scale"&&isNaN(n))return r;if(t==="duration"){let i=n.match(/([0-9]+)ms/);if(i)return i[1]}return t==="origin"&&["top","right","left","center","bottom"].includes(e[e.indexOf(t)+2])?[n,e[e.indexOf(t)+2]].join(" "):n}var Z=!1;function te(e,t=()=>{}){return(...r)=>Z?t(...r):e(...r)}function mn(e){return(...t)=>Z&&e(...t)}function bn(e,t){t._x_dataStack||(t._x_dataStack=e._x_dataStack),Z=!0,En(()=>{wn(t)}),Z=!1}function wn(e){let t=!1;T(e,(n,i)=>{M(n,(o,s)=>{if(t&&ln(o))return s();t=!0,i(o,s)})})}function En(e){let t=k;_t((r,n)=>{let i=t(r);return Q(i),()=>{}}),e(),_t(t)}function er(e,t,r,n=[]){switch(e._x_bindings||(e._x_bindings=H({})),e._x_bindings[t]=r,t=n.includes("camel")?Rn(t):t,t){case"value":Sn(e,r);break;case"style":On(e,r);break;case"class":An(e,r);break;default:Cn(e,t,r);break}}function Sn(e,t){if(e.type==="radio")e.attributes.value===void 0&&(e.value=t),window.fromModel&&(e.checked=vt(e.value,t));else if(e.type==="checkbox")Number.isInteger(t)?e.value=t:!Number.isInteger(t)&&!Array.isArray(t)&&typeof t!="boolean"&&![null,void 0].includes(t)?e.value=String(t):Array.isArray(t)?e.checked=t.some(r=>vt(r,e.value)):e.checked=!!t;else if(e.tagName==="SELECT")Tn(e,t);else{if(e.value===t)return;e.value=t}}function An(e,t){e._x_undoAddedClasses&&e._x_undoAddedClasses(),e._x_undoAddedClasses=ot(e,t)}function On(e,t){e._x_undoAddedStyles&&e._x_undoAddedStyles(),e._x_undoAddedStyles=pe(e,t)}function Cn(e,t,r){[null,void 0,!1].includes(r)&&In(t)?e.removeAttribute(t):(tr(t)&&(r=t),Mn(e,t,r))}function Mn(e,t,r){e.getAttribute(t)!=r&&e.setAttribute(t,r)}function Tn(e,t){let r=[].concat(t).map(n=>n+"");Array.from(e.options).forEach(n=>{n.selected=r.includes(n.value)})}function Rn(e){return e.toLowerCase().replace(/-(\w)/g,(t,r)=>r.toUpperCase())}function vt(e,t){return e==t}function tr(e){return["disabled","checked","required","readonly","hidden","open","selected","autofocus","itemscope","multiple","novalidate","allowfullscreen","allowpaymentrequest","formnovalidate","autoplay","controls","loop","muted","playsinline","default","ismap","reversed","async","defer","nomodule"].includes(e)}function In(e){return!["aria-pressed","aria-checked","aria-expanded","aria-selected"].includes(e)}function $n(e,t,r){if(e._x_bindings&&e._x_bindings[t]!==void 0)return e._x_bindings[t];let n=e.getAttribute(t);return n===null?typeof r=="function"?r():r:n===""?!0:tr(t)?!![t,"true"].includes(n):n}function rr(e,t){var r;return function(){var n=this,i=arguments,o=function(){r=null,e.apply(n,i)};clearTimeout(r),r=setTimeout(o,t)}}function nr(e,t){let r;return function(){let n=this,i=arguments;r||(e.apply(n,i),r=!0,setTimeout(()=>r=!1,t))}}function Pn(e){e(re)}var I={},yt=!1;function Nn(e,t){if(yt||(I=H(I),yt=!0),t===void 0)return I[e];I[e]=t,typeof t=="object"&&t!==null&&t.hasOwnProperty("init")&&typeof t.init=="function"&&I[e].init(),It(I[e])}function jn(){return I}var ir={};function Ln(e,t){let r=typeof t!="function"?()=>t:t;e instanceof Element?or(e,r()):ir[e]=r}function Fn(e){return Object.entries(ir).forEach(([t,r])=>{Object.defineProperty(e,t,{get(){return(...n)=>r(...n)}})}),e}function or(e,t,r){let n=[];for(;n.length;)n.pop()();let i=Object.entries(t).map(([s,a])=>({name:s,value:a})),o=Lt(i);i=i.map(s=>o.find(a=>a.name===s.name)?{name:`x-bind:${s.name}`,value:`"${s.value}"`}:s),et(e,i,r).map(s=>{n.push(s.runCleanups),s()})}var sr={};function Kn(e,t){sr[e]=t}function Dn(e,t){return Object.entries(sr).forEach(([r,n])=>{Object.defineProperty(e,r,{get(){return(...i)=>n.bind(t)(...i)},enumerable:!1})}),e}var Bn={get reactive(){return H},get release(){return Q},get effect(){return k},get raw(){return Et},version:"3.12.0",flushAndStopDeferringMutations:Gr,dontAutoEvaluateFunctions:Qr,disableEffectScheduling:Br,startObservingMutations:Ye,stopObservingMutations:Tt,setReactivityEngine:Hr,closestDataStack:K,skipDuringClone:te,onlyDuringClone:mn,addRootSelector:Jt,addInitSelector:Gt,addScopeToNode:X,deferMutations:Jr,mapAttributes:tt,evaluateLater:b,interceptInit:fn,setEvaluator:Xr,mergeProxies:ee,findClosest:de,closestRoot:fe,destroyTree:Zt,interceptor:$t,transition:ze,setStyles:pe,mutateDom:x,directive:g,throttle:nr,debounce:rr,evaluate:F,initTree:T,nextTick:it,prefixed:z,prefix:nn,plugin:Pn,magic:S,store:Nn,start:cn,clone:bn,bound:$n,$data:Rt,walk:M,data:Kn,bind:Ln},re=Bn;function ar(e,t){let r=Object.create(null),n=e.split(",");for(let i=0;i<n.length;i++)r[n[i]]=!0;return t?i=>!!r[i.toLowerCase()]:i=>!!r[i]}var Hn="itemscope,allowfullscreen,formnovalidate,ismap,nomodule,novalidate,readonly",Ki=ar(Hn+",async,autofocus,autoplay,controls,default,defer,disabled,hidden,loop,open,required,reversed,scoped,seamless,checked,muted,multiple,selected"),kn=Object.freeze({}),Di=Object.freeze([]),ur=Object.assign,zn=Object.prototype.hasOwnProperty,_e=(e,t)=>zn.call(e,t),N=Array.isArray,G=e=>cr(e)==="[object Map]",qn=e=>typeof e=="string",st=e=>typeof e=="symbol",he=e=>e!==null&&typeof e=="object",Wn=Object.prototype.toString,cr=e=>Wn.call(e),lr=e=>cr(e).slice(8,-1),at=e=>qn(e)&&e!=="NaN"&&e[0]!=="-"&&""+parseInt(e,10)===e,ge=e=>{let t=Object.create(null);return r=>t[r]||(t[r]=e(r))},Un=/-(\w)/g,Bi=ge(e=>e.replace(Un,(t,r)=>r?r.toUpperCase():"")),Vn=/\B([A-Z])/g,Hi=ge(e=>e.replace(Vn,"-$1").toLowerCase()),fr=ge(e=>e.charAt(0).toUpperCase()+e.slice(1)),ki=ge(e=>e?`on${fr(e)}`:""),dr=(e,t)=>e!==t&&(e===e||t===t),qe=new WeakMap,W=[],A,j=Symbol("iterate"),We=Symbol("Map key iterate");function Jn(e){return e&&e._isEffect===!0}function Gn(e,t=kn){Jn(e)&&(e=e.raw);let r=Qn(e,t);return t.lazy||r(),r}function Yn(e){e.active&&(pr(e),e.options.onStop&&e.options.onStop(),e.active=!1)}var Zn=0;function Qn(e,t){let r=function(){if(!r.active)return e();if(!W.includes(r)){pr(r);try{return ei(),W.push(r),A=r,e()}finally{W.pop(),_r(),A=W[W.length-1]}}};return r.id=Zn++,r.allowRecurse=!!t.allowRecurse,r._isEffect=!0,r.active=!0,r.raw=e,r.deps=[],r.options=t,r}function pr(e){let{deps:t}=e;if(t.length){for(let r=0;r<t.length;r++)t[r].delete(e);t.length=0}}var B=!0,ut=[];function Xn(){ut.push(B),B=!1}function ei(){ut.push(B),B=!0}function _r(){let e=ut.pop();B=e===void 0?!0:e}function E(e,t,r){if(!B||A===void 0)return;let n=qe.get(e);n||qe.set(e,n=new Map);let i=n.get(r);i||n.set(r,i=new Set),i.has(A)||(i.add(A),A.deps.push(i),A.options.onTrack&&A.options.onTrack({effect:A,target:e,type:t,key:r}))}function R(e,t,r,n,i,o){let s=qe.get(e);if(!s)return;let a=new Set,u=l=>{l&&l.forEach(d=>{(d!==A||d.allowRecurse)&&a.add(d)})};if(t==="clear")s.forEach(u);else if(r==="length"&&N(e))s.forEach((l,d)=>{(d==="length"||d>=n)&&u(l)});else switch(r!==void 0&&u(s.get(r)),t){case"add":N(e)?at(r)&&u(s.get("length")):(u(s.get(j)),G(e)&&u(s.get(We)));break;case"delete":N(e)||(u(s.get(j)),G(e)&&u(s.get(We)));break;case"set":G(e)&&u(s.get(j));break}let c=l=>{l.options.onTrigger&&l.options.onTrigger({effect:l,target:e,key:r,type:t,newValue:n,oldValue:i,oldTarget:o}),l.options.scheduler?l.options.scheduler(l):l()};a.forEach(c)}var ti=ar("__proto__,__v_isRef,__isVue"),hr=new Set(Object.getOwnPropertyNames(Symbol).map(e=>Symbol[e]).filter(st)),ri=ve(),ni=ve(!1,!0),ii=ve(!0),oi=ve(!0,!0),ce={};["includes","indexOf","lastIndexOf"].forEach(e=>{let t=Array.prototype[e];ce[e]=function(...r){let n=h(this);for(let o=0,s=this.length;o<s;o++)E(n,"get",o+"");let i=t.apply(n,r);return i===-1||i===!1?t.apply(n,r.map(h)):i}});["push","pop","shift","unshift","splice"].forEach(e=>{let t=Array.prototype[e];ce[e]=function(...r){Xn();let n=t.apply(this,r);return _r(),n}});function ve(e=!1,t=!1){return function(n,i,o){if(i==="__v_isReactive")return!e;if(i==="__v_isReadonly")return e;if(i==="__v_raw"&&o===(e?t?hi:Tr:t?_i:Mr).get(n))return n;let s=N(n);if(!e&&s&&_e(ce,i))return Reflect.get(ce,i,o);let a=Reflect.get(n,i,o);return(st(i)?hr.has(i):ti(i))||(e||E(n,"get",i),t)?a:Ue(a)?!s||!at(i)?a.value:a:he(a)?e?Rr(a):dt(a):a}}var si=gr(),ai=gr(!0);function gr(e=!1){return function(r,n,i,o){let s=r[n];if(!e&&(i=h(i),s=h(s),!N(r)&&Ue(s)&&!Ue(i)))return s.value=i,!0;let a=N(r)&&at(n)?Number(n)<r.length:_e(r,n),u=Reflect.set(r,n,i,o);return r===h(o)&&(a?dr(i,s)&&R(r,"set",n,i,s):R(r,"add",n,i)),u}}function ui(e,t){let r=_e(e,t),n=e[t],i=Reflect.deleteProperty(e,t);return i&&r&&R(e,"delete",t,void 0,n),i}function ci(e,t){let r=Reflect.has(e,t);return(!st(t)||!hr.has(t))&&E(e,"has",t),r}function li(e){return E(e,"iterate",N(e)?"length":j),Reflect.ownKeys(e)}var vr={get:ri,set:si,deleteProperty:ui,has:ci,ownKeys:li},yr={get:ii,set(e,t){return console.warn(`Set operation on key "${String(t)}" failed: target is readonly.`,e),!0},deleteProperty(e,t){return console.warn(`Delete operation on key "${String(t)}" failed: target is readonly.`,e),!0}},zi=ur({},vr,{get:ni,set:ai}),qi=ur({},yr,{get:oi}),ct=e=>he(e)?dt(e):e,lt=e=>he(e)?Rr(e):e,ft=e=>e,ye=e=>Reflect.getPrototypeOf(e);function xe(e,t,r=!1,n=!1){e=e.__v_raw;let i=h(e),o=h(t);t!==o&&!r&&E(i,"get",t),!r&&E(i,"get",o);let{has:s}=ye(i),a=n?ft:r?lt:ct;if(s.call(i,t))return a(e.get(t));if(s.call(i,o))return a(e.get(o));e!==i&&e.get(t)}function me(e,t=!1){let r=this.__v_raw,n=h(r),i=h(e);return e!==i&&!t&&E(n,"has",e),!t&&E(n,"has",i),e===i?r.has(e):r.has(e)||r.has(i)}function be(e,t=!1){return e=e.__v_raw,!t&&E(h(e),"iterate",j),Reflect.get(e,"size",e)}function xr(e){e=h(e);let t=h(this);return ye(t).has.call(t,e)||(t.add(e),R(t,"add",e,e)),this}function mr(e,t){t=h(t);let r=h(this),{has:n,get:i}=ye(r),o=n.call(r,e);o?Cr(r,n,e):(e=h(e),o=n.call(r,e));let s=i.call(r,e);return r.set(e,t),o?dr(t,s)&&R(r,"set",e,t,s):R(r,"add",e,t),this}function br(e){let t=h(this),{has:r,get:n}=ye(t),i=r.call(t,e);i?Cr(t,r,e):(e=h(e),i=r.call(t,e));let o=n?n.call(t,e):void 0,s=t.delete(e);return i&&R(t,"delete",e,void 0,o),s}function wr(){let e=h(this),t=e.size!==0,r=G(e)?new Map(e):new Set(e),n=e.clear();return t&&R(e,"clear",void 0,void 0,r),n}function we(e,t){return function(n,i){let o=this,s=o.__v_raw,a=h(s),u=t?ft:e?lt:ct;return!e&&E(a,"iterate",j),s.forEach((c,l)=>n.call(i,u(c),u(l),o))}}function oe(e,t,r){return function(...n){let i=this.__v_raw,o=h(i),s=G(o),a=e==="entries"||e===Symbol.iterator&&s,u=e==="keys"&&s,c=i[e](...n),l=r?ft:t?lt:ct;return!t&&E(o,"iterate",u?We:j),{next(){let{value:d,done:p}=c.next();return p?{value:d,done:p}:{value:a?[l(d[0]),l(d[1])]:l(d),done:p}},[Symbol.iterator](){return this}}}}function C(e){return function(...t){{let r=t[0]?`on key "${t[0]}" `:"";console.warn(`${fr(e)} operation ${r}failed: target is readonly.`,h(this))}return e==="delete"?!1:this}}var Er={get(e){return xe(this,e)},get size(){return be(this)},has:me,add:xr,set:mr,delete:br,clear:wr,forEach:we(!1,!1)},Sr={get(e){return xe(this,e,!1,!0)},get size(){return be(this)},has:me,add:xr,set:mr,delete:br,clear:wr,forEach:we(!1,!0)},Ar={get(e){return xe(this,e,!0)},get size(){return be(this,!0)},has(e){return me.call(this,e,!0)},add:C("add"),set:C("set"),delete:C("delete"),clear:C("clear"),forEach:we(!0,!1)},Or={get(e){return xe(this,e,!0,!0)},get size(){return be(this,!0)},has(e){return me.call(this,e,!0)},add:C("add"),set:C("set"),delete:C("delete"),clear:C("clear"),forEach:we(!0,!0)},fi=["keys","values","entries",Symbol.iterator];fi.forEach(e=>{Er[e]=oe(e,!1,!1),Ar[e]=oe(e,!0,!1),Sr[e]=oe(e,!1,!0),Or[e]=oe(e,!0,!0)});function Ee(e,t){let r=t?e?Or:Sr:e?Ar:Er;return(n,i,o)=>i==="__v_isReactive"?!e:i==="__v_isReadonly"?e:i==="__v_raw"?n:Reflect.get(_e(r,i)&&i in n?r:n,i,o)}var di={get:Ee(!1,!1)},Wi={get:Ee(!1,!0)},pi={get:Ee(!0,!1)},Ui={get:Ee(!0,!0)};function Cr(e,t,r){let n=h(r);if(n!==r&&t.call(e,n)){let i=lr(e);console.warn(`Reactive ${i} contains both the raw and reactive versions of the same object${i==="Map"?" as keys":""}, which can lead to inconsistencies. Avoid differentiating between the raw and reactive versions of an object and only use the reactive version if possible.`)}}var Mr=new WeakMap,_i=new WeakMap,Tr=new WeakMap,hi=new WeakMap;function gi(e){switch(e){case"Object":case"Array":return 1;case"Map":case"Set":case"WeakMap":case"WeakSet":return 2;default:return 0}}function vi(e){return e.__v_skip||!Object.isExtensible(e)?0:gi(lr(e))}function dt(e){return e&&e.__v_isReadonly?e:Ir(e,!1,vr,di,Mr)}function Rr(e){return Ir(e,!0,yr,pi,Tr)}function Ir(e,t,r,n,i){if(!he(e))return console.warn(`value cannot be made reactive: ${String(e)}`),e;if(e.__v_raw&&!(t&&e.__v_isReactive))return e;let o=i.get(e);if(o)return o;let s=vi(e);if(s===0)return e;let a=new Proxy(e,s===2?n:r);return i.set(e,a),a}function h(e){return e&&h(e.__v_raw)||e}function Ue(e){return!!(e&&e.__v_isRef===!0)}S("nextTick",()=>it);S("dispatch",e=>J.bind(J,e));S("watch",(e,{evaluateLater:t,effect:r})=>(n,i)=>{let o=t(n),s=!0,a,u=r(()=>o(c=>{JSON.stringify(c),s?a=c:queueMicrotask(()=>{i(c,a),a=c}),s=!1}));e._x_effects.delete(u)});S("store",jn);S("data",e=>Rt(e));S("root",e=>fe(e));S("refs",e=>(e._x_refs_proxy||(e._x_refs_proxy=ee(yi(e))),e._x_refs_proxy));function yi(e){let t=[],r=e;for(;r;)r._x_refs&&t.push(r._x_refs),r=r.parentNode;return t}var Te={};function $r(e){return Te[e]||(Te[e]=0),++Te[e]}function xi(e,t){return de(e,r=>{if(r._x_ids&&r._x_ids[t])return!0})}function mi(e,t){e._x_ids||(e._x_ids={}),e._x_ids[t]||(e._x_ids[t]=$r(t))}S("id",e=>(t,r=null)=>{let n=xi(e,t),i=n?n._x_ids[t]:$r(t);return r?`${t}-${i}-${r}`:`${t}-${i}`});S("el",e=>e);Pr("Focus","focus","focus");Pr("Persist","persist","persist");function Pr(e,t,r){S(t,n=>D(`You can't use [$${directiveName}] without first installing the "${e}" plugin here: https://alpinejs.dev/plugins/${r}`,n))}function bi({get:e,set:t},{get:r,set:n}){let i=!0,o,s,a,u,c=k(()=>{let l,d;i?(l=e(),n(l),d=r(),i=!1):(l=e(),d=r(),a=JSON.stringify(l),u=JSON.stringify(d),a!==o?(d=r(),n(l),d=l):(t(d),l=d)),o=JSON.stringify(l),s=JSON.stringify(d)});return()=>{Q(c)}}g("modelable",(e,{expression:t},{effect:r,evaluateLater:n,cleanup:i})=>{let o=n(t),s=()=>{let l;return o(d=>l=d),l},a=n(`${t} = __placeholder`),u=l=>a(()=>{},{scope:{__placeholder:l}}),c=s();u(c),queueMicrotask(()=>{if(!e._x_model)return;e._x_removeModelListeners.default();let l=e._x_model.get,d=e._x_model.set,p=bi({get(){return l()},set(y){d(y)}},{get(){return s()},set(y){u(y)}});i(p)})});var wi=document.createElement("div");g("teleport",(e,{modifiers:t,expression:r},{cleanup:n})=>{e.tagName.toLowerCase()!=="template"&&D("x-teleport can only be used on a <template> tag",e);let i=te(()=>document.querySelector(r),()=>wi)();i||D(`Cannot find x-teleport element for selector: "${r}"`);let o=e.content.cloneNode(!0).firstElementChild;e._x_teleport=o,o._x_teleportBack=e,e._x_forwardEvents&&e._x_forwardEvents.forEach(s=>{o.addEventListener(s,a=>{a.stopPropagation(),e.dispatchEvent(new a.constructor(a.type,a))})}),X(o,{},e),x(()=>{t.includes("prepend")?i.parentNode.insertBefore(o,i):t.includes("append")?i.parentNode.insertBefore(o,i.nextSibling):i.appendChild(o),T(o),o._x_ignore=!0}),n(()=>o.remove())});var Nr=()=>{};Nr.inline=(e,{modifiers:t},{cleanup:r})=>{t.includes("self")?e._x_ignoreSelf=!0:e._x_ignore=!0,r(()=>{t.includes("self")?delete e._x_ignoreSelf:delete e._x_ignore})};g("ignore",Nr);g("effect",(e,{expression:t},{effect:r})=>r(b(e,t)));function Ve(e,t,r,n){let i=e,o=u=>n(u),s={},a=(u,c)=>l=>c(u,l);if(r.includes("dot")&&(t=Ei(t)),r.includes("camel")&&(t=Si(t)),r.includes("passive")&&(s.passive=!0),r.includes("capture")&&(s.capture=!0),r.includes("window")&&(i=window),r.includes("document")&&(i=document),r.includes("prevent")&&(o=a(o,(u,c)=>{c.preventDefault(),u(c)})),r.includes("stop")&&(o=a(o,(u,c)=>{c.stopPropagation(),u(c)})),r.includes("self")&&(o=a(o,(u,c)=>{c.target===e&&u(c)})),(r.includes("away")||r.includes("outside"))&&(i=document,o=a(o,(u,c)=>{e.contains(c.target)||c.target.isConnected!==!1&&(e.offsetWidth<1&&e.offsetHeight<1||e._x_isShown!==!1&&u(c))})),r.includes("once")&&(o=a(o,(u,c)=>{u(c),i.removeEventListener(t,o,s)})),o=a(o,(u,c)=>{Oi(t)&&Ci(c,r)||u(c)}),r.includes("debounce")){let u=r[r.indexOf("debounce")+1]||"invalid-wait",c=le(u.split("ms")[0])?Number(u.split("ms")[0]):250;o=rr(o,c)}if(r.includes("throttle")){let u=r[r.indexOf("throttle")+1]||"invalid-wait",c=le(u.split("ms")[0])?Number(u.split("ms")[0]):250;o=nr(o,c)}return i.addEventListener(t,o,s),()=>{i.removeEventListener(t,o,s)}}function Ei(e){return e.replace(/-/g,".")}function Si(e){return e.toLowerCase().replace(/-(\w)/g,(t,r)=>r.toUpperCase())}function le(e){return!Array.isArray(e)&&!isNaN(e)}function Ai(e){return[" ","_"].includes(e)?e:e.replace(/([a-z])([A-Z])/g,"$1-$2").replace(/[_\s]/,"-").toLowerCase()}function Oi(e){return["keydown","keyup"].includes(e)}function Ci(e,t){let r=t.filter(o=>!["window","document","prevent","stop","once","capture"].includes(o));if(r.includes("debounce")){let o=r.indexOf("debounce");r.splice(o,le((r[o+1]||"invalid-wait").split("ms")[0])?2:1)}if(r.includes("throttle")){let o=r.indexOf("throttle");r.splice(o,le((r[o+1]||"invalid-wait").split("ms")[0])?2:1)}if(r.length===0||r.length===1&&xt(e.key).includes(r[0]))return!1;let i=["ctrl","shift","alt","meta","cmd","super"].filter(o=>r.includes(o));return r=r.filter(o=>!i.includes(o)),!(i.length>0&&i.filter(s=>((s==="cmd"||s==="super")&&(s="meta"),e[`${s}Key`])).length===i.length&&xt(e.key).includes(r[0]))}function xt(e){if(!e)return[];e=Ai(e);let t={ctrl:"control",slash:"/",space:" ",spacebar:" ",cmd:"meta",esc:"escape",up:"arrow-up",down:"arrow-down",left:"arrow-left",right:"arrow-right",period:".",equal:"=",minus:"-",underscore:"_"};return t[e]=e,Object.keys(t).map(r=>{if(t[r]===e)return r}).filter(r=>r)}g("model",(e,{modifiers:t,expression:r},{effect:n,cleanup:i})=>{let o=e;t.includes("parent")&&(o=e.parentNode);let s=b(o,r),a;typeof r=="string"?a=b(o,`${r} = __placeholder`):typeof r=="function"&&typeof r()=="string"?a=b(o,`${r()} = __placeholder`):a=()=>{};let u=()=>{let p;return s(y=>p=y),mt(p)?p.get():p},c=p=>{let y;s(O=>y=O),mt(y)?y.set(p):a(()=>{},{scope:{__placeholder:p}})};t.includes("fill")&&e.hasAttribute("value")&&(u()===null||u()==="")&&c(e.value),typeof r=="string"&&e.type==="radio"&&x(()=>{e.hasAttribute("name")||e.setAttribute("name",r)});var l=e.tagName.toLowerCase()==="select"||["checkbox","radio"].includes(e.type)||t.includes("lazy")?"change":"input";let d=Z?()=>{}:Ve(e,l,t,p=>{c(Mi(e,t,p,u()))});if(e._x_removeModelListeners||(e._x_removeModelListeners={}),e._x_removeModelListeners.default=d,i(()=>e._x_removeModelListeners.default()),e.form){let p=Ve(e.form,"reset",[],y=>{it(()=>e._x_model&&e._x_model.set(e.value))});i(()=>p())}e._x_model={get(){return u()},set(p){c(p)}},e._x_forceModelUpdate=p=>{p=p===void 0?u():p,p===void 0&&typeof r=="string"&&r.match(/\./)&&(p=""),window.fromModel=!0,x(()=>er(e,"value",p)),delete window.fromModel},n(()=>{let p=u();t.includes("unintrusive")&&document.activeElement.isSameNode(e)||e._x_forceModelUpdate(p)})});function Mi(e,t,r,n){return x(()=>{if(r instanceof CustomEvent&&r.detail!==void 0)return typeof r.detail<"u"?r.detail:r.target.value;if(e.type==="checkbox")if(Array.isArray(n)){let i=t.includes("number")?Re(r.target.value):r.target.value;return r.target.checked?n.concat([i]):n.filter(o=>!Ti(o,i))}else return r.target.checked;else{if(e.tagName.toLowerCase()==="select"&&e.multiple)return t.includes("number")?Array.from(r.target.selectedOptions).map(i=>{let o=i.value||i.text;return Re(o)}):Array.from(r.target.selectedOptions).map(i=>i.value||i.text);{let i=r.target.value;return t.includes("number")?Re(i):t.includes("trim")?i.trim():i}}})}function Re(e){let t=e?parseFloat(e):null;return Ri(t)?t:e}function Ti(e,t){return e==t}function Ri(e){return!Array.isArray(e)&&!isNaN(e)}function mt(e){return e!==null&&typeof e=="object"&&typeof e.get=="function"&&typeof e.set=="function"}g("cloak",e=>queueMicrotask(()=>x(()=>e.removeAttribute(z("cloak")))));Gt(()=>`[${z("init")}]`);g("init",te((e,{expression:t},{evaluate:r})=>typeof t=="string"?!!t.trim()&&r(t,{},!1):r(t,{},!1)));g("text",(e,{expression:t},{effect:r,evaluateLater:n})=>{let i=n(t);r(()=>{i(o=>{x(()=>{e.textContent=o})})})});g("html",(e,{expression:t},{effect:r,evaluateLater:n})=>{let i=n(t);r(()=>{i(o=>{x(()=>{e.innerHTML=o,e._x_ignoreSelf=!0,T(e),delete e._x_ignoreSelf})})})});tt(Dt(":",Bt(z("bind:"))));g("bind",(e,{value:t,modifiers:r,expression:n,original:i},{effect:o})=>{if(!t){let a={};Fn(a),b(e,n)(c=>{or(e,c,i)},{scope:a});return}if(t==="key")return Ii(e,n);let s=b(e,n);o(()=>s(a=>{a===void 0&&typeof n=="string"&&n.match(/\./)&&(a=""),x(()=>er(e,t,a,r))}))});function Ii(e,t){e._x_keyExpression=t}Jt(()=>`[${z("data")}]`);g("data",te((e,{expression:t},{cleanup:r})=>{t=t===""?"{}":t;let n={};Le(n,e);let i={};Dn(i,n);let o=F(e,t,{scope:i});(o===void 0||o===!0)&&(o={}),Le(o,e);let s=H(o);It(s);let a=X(e,s);s.init&&F(e,s.init),r(()=>{s.destroy&&F(e,s.destroy),a()})}));g("show",(e,{modifiers:t,expression:r},{effect:n})=>{let i=b(e,r);e._x_doHide||(e._x_doHide=()=>{x(()=>{e.style.setProperty("display","none",t.includes("important")?"important":void 0)})}),e._x_doShow||(e._x_doShow=()=>{x(()=>{e.style.length===1&&e.style.display==="none"?e.removeAttribute("style"):e.style.removeProperty("display")})});let o=()=>{e._x_doHide(),e._x_isShown=!1},s=()=>{e._x_doShow(),e._x_isShown=!0},a=()=>setTimeout(s),u=ke(d=>d?s():o(),d=>{typeof e._x_toggleAndCascadeWithTransitions=="function"?e._x_toggleAndCascadeWithTransitions(e,d,s,o):d?a():o()}),c,l=!0;n(()=>i(d=>{!l&&d===c||(t.includes("immediate")&&(d?a():o()),u(d),c=d,l=!1)}))});g("for",(e,{expression:t},{effect:r,cleanup:n})=>{let i=Pi(t),o=b(e,i.items),s=b(e,e._x_keyExpression||"index");e._x_prevKeys=[],e._x_lookup={},r(()=>$i(e,i,o,s)),n(()=>{Object.values(e._x_lookup).forEach(a=>a.remove()),delete e._x_prevKeys,delete e._x_lookup})});function $i(e,t,r,n){let i=s=>typeof s=="object"&&!Array.isArray(s),o=e;r(s=>{Ni(s)&&s>=0&&(s=Array.from(Array(s).keys(),f=>f+1)),s===void 0&&(s=[]);let a=e._x_lookup,u=e._x_prevKeys,c=[],l=[];if(i(s))s=Object.entries(s).map(([f,_])=>{let v=bt(t,_,f,s);n(m=>l.push(m),{scope:{index:f,...v}}),c.push(v)});else for(let f=0;f<s.length;f++){let _=bt(t,s[f],f,s);n(v=>l.push(v),{scope:{index:f,..._}}),c.push(_)}let d=[],p=[],y=[],O=[];for(let f=0;f<u.length;f++){let _=u[f];l.indexOf(_)===-1&&y.push(_)}u=u.filter(f=>!y.includes(f));let ne="template";for(let f=0;f<l.length;f++){let _=l[f],v=u.indexOf(_);if(v===-1)u.splice(f,0,_),d.push([ne,f]);else if(v!==f){let m=u.splice(f,1)[0],w=u.splice(v-1,1)[0];u.splice(f,0,w),u.splice(v,0,m),p.push([m,w])}else O.push(_);ne=_}for(let f=0;f<y.length;f++){let _=y[f];a[_]._x_effects&&a[_]._x_effects.forEach(wt),a[_].remove(),a[_]=null,delete a[_]}for(let f=0;f<p.length;f++){let[_,v]=p[f],m=a[_],w=a[v],L=document.createElement("div");x(()=>{w.after(L),m.after(w),w._x_currentIfEl&&w.after(w._x_currentIfEl),L.before(m),m._x_currentIfEl&&m.after(m._x_currentIfEl),L.remove()}),ht(w,c[l.indexOf(v)])}for(let f=0;f<d.length;f++){let[_,v]=d[f],m=_==="template"?o:a[_];m._x_currentIfEl&&(m=m._x_currentIfEl);let w=c[v],L=l[v],ie=document.importNode(o.content,!0).firstElementChild;X(ie,H(w),o),x(()=>{m.after(ie),T(ie)}),typeof L=="object"&&D("x-for key cannot be an object, it must be a string or an integer",o),a[L]=ie}for(let f=0;f<O.length;f++)ht(a[O[f]],c[l.indexOf(O[f])]);o._x_prevKeys=l})}function Pi(e){let t=/,([^,\}\]]*)(?:,([^,\}\]]*))?$/,r=/^\s*\(|\)\s*$/g,n=/([\s\S]*?)\s+(?:in|of)\s+([\s\S]*)/,i=e.match(n);if(!i)return;let o={};o.items=i[2].trim();let s=i[1].replace(r,"").trim(),a=s.match(t);return a?(o.item=s.replace(t,"").trim(),o.index=a[1].trim(),a[2]&&(o.collection=a[2].trim())):o.item=s,o}function bt(e,t,r,n){let i={};return/^\[.*\]$/.test(e.item)&&Array.isArray(t)?e.item.replace("[","").replace("]","").split(",").map(s=>s.trim()).forEach((s,a)=>{i[s]=t[a]}):/^\{.*\}$/.test(e.item)&&!Array.isArray(t)&&typeof t=="object"?e.item.replace("{","").replace("}","").split(",").map(s=>s.trim()).forEach(s=>{i[s]=t[s]}):i[e.item]=t,e.index&&(i[e.index]=r),e.collection&&(i[e.collection]=n),i}function Ni(e){return!Array.isArray(e)&&!isNaN(e)}function jr(){}jr.inline=(e,{expression:t},{cleanup:r})=>{let n=fe(e);n._x_refs||(n._x_refs={}),n._x_refs[t]=e,r(()=>delete n._x_refs[t])};g("ref",jr);g("if",(e,{expression:t},{effect:r,cleanup:n})=>{let i=b(e,t),o=()=>{if(e._x_currentIfEl)return e._x_currentIfEl;let a=e.content.cloneNode(!0).firstElementChild;return X(a,{},e),x(()=>{e.after(a),T(a)}),e._x_currentIfEl=a,e._x_undoIf=()=>{M(a,u=>{u._x_effects&&u._x_effects.forEach(wt)}),a.remove(),delete e._x_currentIfEl},a},s=()=>{e._x_undoIf&&(e._x_undoIf(),delete e._x_undoIf)};r(()=>i(a=>{a?o():s()})),n(()=>e._x_undoIf&&e._x_undoIf())});g("id",(e,{expression:t},{evaluate:r})=>{r(t).forEach(i=>mi(e,i))});tt(Dt("@",Bt(z("on:"))));g("on",te((e,{value:t,modifiers:r,expression:n},{cleanup:i})=>{let o=n?b(e,n):()=>{};e.tagName.toLowerCase()==="template"&&(e._x_forwardEvents||(e._x_forwardEvents=[]),e._x_forwardEvents.includes(t)||e._x_forwardEvents.push(t));let s=Ve(e,t,r,a=>{o(()=>{},{scope:{$event:a},params:[a]})});i(()=>s())}));Se("Collapse","collapse","collapse");Se("Intersect","intersect","intersect");Se("Focus","trap","focus");Se("Mask","mask","mask");function Se(e,t,r){g(t,n=>D(`You can't use [x-${t}] without first installing the "${e}" plugin here: https://alpinejs.dev/plugins/${r}`,n))}re.setEvaluator(jt);re.setReactivityEngine({reactive:dt,effect:Gn,release:Yn,raw:h});var ji=re,pt=ji;var Ae=Date.now,Oe=window.requestAnimationFrame,Li=(e,t)=>{let r=Ae(),n=!1,i=()=>{Ae()-r<t||(r+=t,e()),n||Oe(i)};return Oe(i),{clear:()=>n=!0}},Fi=(e,t)=>{let r=Ae(),n=!1,i=()=>{Ae()-r<t?n||Oe(i):e()};return Oe(i),{clear:()=>n=!0}};pt.data("health",()=>({cssClasses:{OK:"bg-green-700",WARNING:"bg-yellow-700",ERROR:"bg-red-700",UNKNOWN:"bg-[#323232]",NOT_RUN:"bg-[#323232]"},checks:{},canCopy:window.navigator.clipboard!==void 0,copiedCode:!1,copyTimeout:null,copy(){this.copiedCode=[...this.$el.querySelectorAll("code")].map(e=>e.innerText).join(`
+`),window.navigator.clipboard.writeText(this.copiedCode),this.copyTimeout&&this.copyTimeout.clear(),this.copyTimeout=Fi(()=>{this.copiedCode=!1},5e3)},fetch(e){fetch(`/setup/${e}.json`).then(t=>!t.ok&&t.status!==503?{errorResponse:{status:"ERROR",title:"Flow Framework",message:"Flow didn't respond as expected."}}:t.json()).then(t=>{this.checks={...this.checks,...t}})},init(){this.fetch("compiletime"),this.fetch("runtime"),Li(()=>{this.fetch("compiletime"),this.fetch("runtime")},5e3)}}));pt.start();})();
